+++ shoop/core/models/_orders.py
@@ -11,6 +11,7 @@ import datetime
 from collections import defaultdict
 from decimal import Decimal
 
+import six
 from django.conf import settings
 from django.core.exceptions import ObjectDoesNotExist, ValidationError
 from django.db import models
@@ -161,6 +162,7 @@ class Order(MoneyPropped, models.Model):
     # Identification
     shop = UnsavedForeignKey("Shop", on_delete=models.PROTECT, verbose_name=_('shop'))
     created_on = models.DateTimeField(auto_now_add=True, editable=False, verbose_name=_('created on'))
+    modified_on = models.DateTimeField(auto_now_add=True, editable=False, verbose_name=_('modified on'))
     identifier = InternalIdentifierField(unique=True, db_index=True, verbose_name=_('order identifier'))
     # TODO: label is actually a choice field, need to check migrations/choice deconstruction
     label = models.CharField(max_length=32, db_index=True, verbose_name=_('label'))
@@ -199,6 +201,10 @@ class Order(MoneyPropped, models.Model):
         settings.AUTH_USER_MODEL, related_name='orders_created', blank=True, null=True,
         on_delete=models.PROTECT,
         verbose_name=_('creating user'))
+    modified_by = UnsavedForeignKey(
+        settings.AUTH_USER_MODEL, related_name='orders_modified', blank=True, null=True,
+        on_delete=models.PROTECT,
+        verbose_name=_('modifier user'))
     deleted = models.BooleanField(db_index=True, default=False, verbose_name=_('deleted'))
     status = UnsavedForeignKey("OrderStatus", verbose_name=_('status'), on_delete=models.PROTECT)
     payment_status = EnumIntegerField(
@@ -255,6 +261,7 @@ class Order(MoneyPropped, models.Model):
     require_verification = models.BooleanField(default=False, verbose_name=_('requires verification'))
     all_verified = models.BooleanField(default=False, verbose_name=_('all lines verified'))
     marketing_permission = models.BooleanField(default=True, verbose_name=_('marketing permission'))
+    _codes = JSONField(blank=True, null=True, verbose_name=_('codes'))
 
     common_select_related = ("billing_address",)
     objects = OrderQuerySet.as_manager()
@@ -273,6 +280,19 @@ class Order(MoneyPropped, models.Model):
             return "Order %s (%s, %s)" % (self.identifier, self.shop.name, name)
         else:
             return "Order %s (%s)" % (self.identifier, name)
+
+    @property
+    def codes(self):
+        return list(self._codes or [])
+
+    @codes.setter
+    def codes(self, value):
+        codes = []
+        for code in value:
+            if not isinstance(code, six.text_type):
+                raise TypeError('codes must be a list of strings')
+            codes.append(code)
+        self._codes = codes
 
     def cache_prices(self):
         taxful_total = TaxfulPrice(0, self.currency)
@@ -328,6 +348,9 @@ class Order(MoneyPropped, models.Model):
 
         if not self.key:
             self.key = get_random_string(32)
+
+        if not self.modified_by:
+            self.modified_by = self.creator
 
     def _save_identifiers(self):
         self.identifier = "%s" % (get_order_identifier(self))