+++ shoop/core/order_creator/_source.py
@@ -113,6 +113,7 @@ class OrderSource(object):
         self._customer = None
         self._orderer = None
         self._creator = None
+        self._modified_by = None
         self.shipping_method_id = None
         self.payment_method_id = None
         self.customer_comment = u""
@@ -153,6 +154,7 @@ class OrderSource(object):
             customer=order.customer,
             orderer=order.orderer,
             creator=order.creator,
+            modified_by=order.modified_by,
             payment_method_id=order.payment_method_id,
             shipping_method_id=order.shipping_method_id,
             customer_comment=order.customer_comment,
@@ -166,6 +168,7 @@ class OrderSource(object):
             payment_data=order.payment_data,
             shipping_data=order.shipping_data,
             extra_data=order.extra_data,
+            codes=order.codes
         )
 
     total_price = _PriceSum("price")
@@ -205,6 +208,14 @@ class OrderSource(object):
     @creator.setter
     def creator(self, value):
         self._creator = value
+
+    @property
+    def modified_by(self):
+        return (self._modified_by or self.creator)
+
+    @modified_by.setter
+    def modified_by(self, value):
+        self._modified_by = value
 
     @property
     def shipping_method(self):
@@ -250,6 +261,11 @@ class OrderSource(object):
     @property
     def codes(self):
         return list(self._codes)
+
+    @codes.setter
+    def codes(self, value):
+        for code in value:
+            self.add_code(code)
 
     def add_code(self, code):
         """