+++ shoop_tests/admin/test_order_creator.py
@@ -22,7 +22,7 @@ from shoop.default_tax.models import TaxRule
 from shoop.testing.factories import (
     create_empty_order, create_product, create_random_company, create_random_person,
     get_default_payment_method, get_default_shipping_method, get_default_shop,
-    get_default_supplier, get_initial_order_status
+    get_default_supplier, get_initial_order_status, UserFactory
 )
 from shoop.testing.utils import apply_request_middleware
 from shoop_tests.utils import assert_contains, printable_gibberish
@@ -259,15 +259,20 @@ def test_company_contact_creation(rf, admin_user):
 
 
 def test_editing_existing_order(rf, admin_user):
+    modifier = UserFactory()
     get_initial_order_status()  # Needed for the API
     contact = create_random_person(locale="en_US", minimum_name_comp_len=5)
     state = get_frontend_order_state(contact=contact)
     shop = get_default_shop()
-    order = create_empty_order(shop)
+    order = create_empty_order(shop=shop)
+    order.payment_data = {"payment_data": True}
+    order.shipping_data = {"shipping_data": True}
+    order.extra_data = {"external_id": "123"}
     order.save()
     assert order.lines.count() == 0
     assert order.pk is not None
-    request = get_frontend_request_for_command(state, "finalize", admin_user)
+    assert order.modified_by == order.creator
+    request = get_frontend_request_for_command(state, "finalize", modifier)
     response = OrderEditView.as_view()(request, pk=order.pk)
     assert_contains(response, "orderIdentifier")  # this checks for status codes as a side effect
     data = json.loads(response.content.decode("utf8"))
@@ -279,7 +284,6 @@ def test_editing_existing_order(rf, admin_user):
 
     # Check that the product content is updated based on state
     assert edited_order.lines.count() == 5
-    assert edited_order.creator == admin_user
     assert edited_order.customer == contact
 
     # Check that product line have right taxes
@@ -287,3 +291,20 @@ def test_editing_existing_order(rf, admin_user):
         if line.type == OrderLineType.PRODUCT:
             assert [line_tax.tax.code for line_tax in line.taxes.all()] == ["test_code"]
             assert line.taxful_price.amount > line.taxless_price.amount
+
+    # Make sure order modification information is correct
+    assert edited_order.modified_by != order.modified_by
+    assert edited_order.modified_by == modifier
+    assert edited_order.modified_on > order.modified_on
+
+    # Make sure all non handled attributes is preserved from original order
+    assert edited_order.creator == order.creator
+    assert edited_order.ip_address == order.ip_address
+    assert edited_order.orderer == order.orderer
+    assert edited_order.customer_comment == order.customer_comment
+    assert edited_order.marketing_permission == order.marketing_permission
+    assert edited_order.order_date == order.order_date
+    assert edited_order.status == order.status
+    assert edited_order.payment_data == order.payment_data
+    assert edited_order.shipping_data == order.shipping_data
+    assert edited_order.extra_data == order.extra_data